refactor: moved the connections section to a new file
