+++ components/helpers/user-button/security/security-section.tsx
@@ -1,6 +1,8 @@
 import { User } from "@prisma/client";
 import { PasswordSection } from "./password-section";
 import { TwoFactorSection } from "./two-factor-section";
+import { PasskeySection } from "./passkey-section";
+import { ConnectionsSection } from "./connections-section";
 
 export const SecuritySection = ({
   user,
@@ -12,249 +14,9 @@ export const SecuritySection = ({
       <p className="text-sm font-medium pointer-events-none">Security</p>
       <div className="flex w-full md:items-end flex-col gap-10">
         <PasswordSection />
-
         <TwoFactorSection user={user} />
-
-
-
-        <div className="flex md:w-[72%] flex-col gap-10">
-          <div
-            ref={thirdParent}
-            className="flex md:flex-row flex-col md:gap-0 gap-8 justify-between"
-          >
-            <p className="text-sm font-medium">Connected Accounts</p>
-            <div
-              ref={thirdParent}
-              className="flex flex-col gap-1 md:items-end md:w-[350px] md:ml-0 ml-4"
-            >
-              {error && (
-                <ErrorCard className="w-full -mt-2" size="sm" error={error} />
-              )}
-              {connections.map((connection) => {
-                // @ts-expect-error Just a simple type error
-                const provider: string = connection.provider;
-                const formattedProvider =
-                  provider.charAt(0).toUpperCase() + provider.slice(1);
-
-                return (
-                  <div
-                    key={connection.id}
-                    ref={thirdParent}
-                    className="min-w-[350px] -mt-1"
-                  >
-                    <div className="flex items-center gap-4">
-                      <div
-                        key={connection.id}
-                        className={cn(
-                          "flex items-center justify-between -mt-2 gap-4",
-                          provider === "credential" && "hidden"
-                        )}
-                      >
-                        <div className={cn("flex items-center gap-2")}>
-                          {provider === "google" && <FcGoogle size={18} />}
-                          {provider === "github" && <FaGithub size={18} />}
-                          <p className="text-sm">{formattedProvider}</p>
-                          <p className="text-sm text-zinc-500">•</p>
-                          <p className="text-sm text-zinc-500/85">
-                            {user.email}
-                          </p>
-                        </div>
-                      </div>
-                      <DropdownMenu>
-                        <DropdownMenuTrigger
-                          asChild
-                          className={cn(
-                            "inline-flex ml-auto",
-                            provider === "credential" && "hidden"
-                          )}
-                        >
-                          <Button
-                            variant={"ghost"}
-                            size="icon"
-                            className="group"
-                          >
-                            <Ellipsis className="h-4 w-4 text-zinc-400 group-hover:text-zinc-800 transition" />
-                          </Button>
-                        </DropdownMenuTrigger>
-                        <DropdownMenuContent
-                          align="center"
-                          className="min-w-fit py-0 px-0"
-                        >
-                          <DropdownMenuItem
-                            className="cursor-pointer px-3 py-1 text-zinc-600 focus:text-zinc-800 transition-all"
-                            onClick={() => {
-                              setIsDeleteConnectionBoxOpen(
-                                provider === "github" ? "github" : "google"
-                              );
-                            }}
-                          >
-                            <p className="text-sm text-destructive">Remove</p>
-                          </DropdownMenuItem>
-                        </DropdownMenuContent>
-                      </DropdownMenu>
-                    </div>
-                    {isDeleteConnectionBoxOpen === "github" && (
-                      <Card
-                        className={cn(
-                          "my-4 shadow-md w-full bg-muted-foreground/5 border-zinc-600/15 md:max-w-[350px]",
-                          isDeleteConnectionBoxOpen === "github" &&
-                            provider !== "github" &&
-                            "hidden"
-                        )}
-                      >
-                        <CardHeader className="w-full flex flex-col">
-                          <CardTitle className="text-sm tracking-tight">
-                            Remove Connection
-                          </CardTitle>
-                          <CardDescription className="text-xs">
-                            Are you sure you want to remove this connection?
-                            <br />
-                            This action cannot be undone.
-                          </CardDescription>
-                        </CardHeader>
-                        <CardContent>
-                          <Button
-                            disabled={isLoading}
-                            onClick={() =>
-                              setIsDeleteConnectionBoxOpen("closed")
-                            }
-                            variant={"ghost"}
-                            size={"sm"}
-                            className="mt-4 mr-2"
-                          >
-                            Cancel
-                          </Button>
-                          <Button
-                            ref={thirdParent}
-                            onClick={() => onGithubDelete()}
-                            variant={"destructive"}
-                            disabled={isLoading}
-                            size={"sm"}
-                            className="mt-4 mr-2"
-                          >
-                            {isLoading && (
-                              <Loader className="size-2 text-white animate-spin" />
-                            )}
-                            {!isLoading && "Delete"}
-                          </Button>
-                        </CardContent>
-                      </Card>
-                    )}
-                    {isDeleteConnectionBoxOpen === "google" && (
-                      <Card
-                        className={cn(
-                          "my-4 shadow-md w-full bg-muted-foreground/5 border-zinc-600/15 md:max-w-[350px]",
-                          isDeleteConnectionBoxOpen === "google" &&
-                            provider !== "google" &&
-                            "hidden"
-                        )}
-                      >
-                        <CardHeader className="w-full flex flex-col">
-                          <CardTitle className="text-sm tracking-tight">
-                            Remove Connection
-                          </CardTitle>
-                          <CardDescription className="text-xs">
-                            Are you sure you want to remove this connection?
-                            <br />
-                            This action cannot be undone.
-                          </CardDescription>
-                        </CardHeader>
-                        <CardContent>
-                          <Button
-                            disabled={isLoading}
-                            onClick={() =>
-                              setIsDeleteConnectionBoxOpen("closed")
-                            }
-                            variant={"ghost"}
-                            size={"sm"}
-                            className="mt-4 mr-2"
-                          >
-                            Cancel
-                          </Button>
-                          <Button
-                            ref={thirdParent}
-                            onClick={onGoogleDelete}
-                            variant={"destructive"}
-                            disabled={isLoading}
-                            size={"sm"}
-                            className="mt-4 mr-2"
-                          >
-                            {isLoading && (
-                              <Loader className="size-2 text-white animate-spin" />
-                            )}
-                            {!isLoading && "Delete"}
-                          </Button>
-                        </CardContent>
-                      </Card>
-                    )}
-                  </div>
-                );
-              })}
-              <DropdownMenu>
-                <DropdownMenuTrigger asChild>
-                  <Button
-                    variant={"ghost"}
-                    size="sm"
-                    className={cn(
-                      "text-sm self-start -mt-[7.5px]",
-                      // @ts-expect-error Just a simple type error
-                      passkeys.data?.length > 0 ? "-ml-3" : "md:ml-auto"
-                    )}
-                    effect={"expandIcon"}
-                    icon={ArrowRight}
-                    iconPlacement="right"
-                  >
-                    Add connection
-                  </Button>
-                </DropdownMenuTrigger>
-                <DropdownMenuContent
-                  align="start"
-                  className="rounded-lg shadow-lg min-w-[180px] py-[0.5px]"
-                >
-                  <DropdownMenuItem
-                    className="cursor-pointer"
-                    onClick={async () => {
-                      await authClient.linkSocial(
-                        {
-                          provider: "google",
-                          callbackURL: "/profile",
-                        },
-                        {
-                          onError: (ctx) => {
-                            setError(ctx.error.message);
-                          },
-                        }
-                      );
-                    }}
-                  >
-                    <FcGoogle size={18} />
-                    <p className="text-sm text-zinc-600">Google</p>
-                  </DropdownMenuItem>
-                  <DropdownMenuSeparator />
-                  <DropdownMenuItem
-                    className="cursor-pointer"
-                    onClick={async () => {
-                      await authClient.linkSocial(
-                        {
-                          provider: "github",
-                          callbackURL: "/profile",
-                        },
-                        {
-                          onError: (ctx) => {
-                            setError(ctx.error.message);
-                          },
-                        }
-                      );
-                    }}
-                  >
-                    <FaGithub size={18} />
-                    <p className="text-sm text-zinc-600">Github</p>
-                  </DropdownMenuItem>
-                </DropdownMenuContent>
-              </DropdownMenu>
-            </div>
-          </div>
-        </div>
+        <PasskeySection />
+        <ConnectionsSection user={user} />
       </div>
     </div>
   );
